requirements.txt
BTrees==4.9.2
cffi==1.14.6
click==8.0.1
Flask==2.0.1
greenlet==1.1.1
gunicorn==20.1.0
inflect==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
persistent==4.7.0
Pillow==8.3.2
pycparser==2.20
pycurl==7.44.1
pymongo==3.12.0
six==1.16.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.23
transaction==3.0.1
Werkzeug==2.0.1
zc.lockfile==2.0
ZConfig==3.6.0
ZODB==5.6.0
zodbpickle==2.0.0
zope.component==5.0.1
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
